power: Use mock path for 'sync_on_suspend' in unit tests
When key eviction is enabled powerd tries to write to
/sys/power/sync_on_suspend during suspend to disable
syncing. The DaemonTest.PrepareToSuspendAndResume unit
test currently fails with USE=key_eviction because the
test host doesn't have permissions to write to the
sysfs of the test host.

Use a mock path for unit testing, as we already do in
other similar cases.

BUG=b:333919993
TEST=USE=key_eviction FEATURES=test \
  P2_TEST_FILTER="DaemonTest.PrepareToSuspendAndResume" \
  emerge-brya power_manager

Change-Id: If8e812aa1ec0245095712c305b5b8c90f040d720
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5444021
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>