patchpanel: Move IPv4 local route to per-device table
When an IP address is configured with a prefix, Linux kernel will add a
local route for that prefix to the network interface automatically. In
the case of IPv4 these routes are added to main table, and in the case
IPv6 those are added to the per-device table configured by
accept_ra_rt_table. We currently do not touch those routes.

However, the current behavior has two problems. First, when an IPv6
network is connected as non-primary, even the local network cannot be
reached as the route to primary network will take precedence
(b/235050937). Second, when the primary and secondary networks are
connected to a same IPv4 network (multihoming), the device local traffic
will be routed to is non-deterministic, as both routes lie in the main
table with a same metric (b/329460927).

This patch fixes the issues by moving the local route to per-device
table after we add an IPv4 address, and add a destination-matching rule
to ensure those get queried according to the priority.

Sample routing policy table after the patch:
1000:  from all lookup main
1010:  from all to 192.168.0.0/24 lookup 1002
1010:  from 192.168.0.100 lookup 1002
1010:  from all fwmark 0x3ea0000/0xffff0000 lookup 1002
1010:  from all oif eth0 lookup 1002
1010:  from all iif eth0 lookup 1002
1020:  from all to 192.168.0.0.24 lookup 1003
1020:  from 192.168.0.200 lookup 1003
1020:  from all fwmark 0x3eb0000/0xffff0000 lookup 1003
1020:  from all oif wlan0 lookup 1003
1020:  from all iif wlan0 lookup 1003
(...)

BUG=b:235050937
BUG=b:329460927
TEST=unit
TEST=tast run $DUT network.Routing*
TEST=tast run $DUT network.VPNRouting*
TEST=tast run $DUT network.ARCIPv6Connectivity*

Change-Id: Icdfab3c31ea8614b778c4a928a2cbb181c6f1da5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5447212
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Taoyu Li <taoyl@chromium.org>
Tested-by: Taoyu Li <taoyl@chromium.org>