typecd: Add unit tests for the DBusManager
The DBusManager in typecd is growing to include informational methods
which other processes can use to get the status of USB-C ports, partners
and cables. Currently, typecd does not have any unit tests for the
DBusManager. Add a set of unit tests checking the proper functionality
of the DBusManager.

BUG=b:334104230
TEST=FEATURES=test emerge-${BOARD} chromeos-base/typecd

Change-Id: I1644b9c511bebe7e497dc60e4439363e23e46df3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5450110
Tested-by: Jameson Thies <jthies@google.com>
Commit-Queue: Jameson Thies <jthies@google.com>
Reviewed-by: Bartosz Szpila <bszpila@google.com>
Reviewed-by: Benson Leung <bleung@google.com>