login: Check for flex_id_tool rather than HWID
Checking for a tool is more robust than checking for just a single
HWID. If flex_id_tool is found in $PATH, then use it to generate a
flex_id entry. Otherwise, we're not on Flex and we generate a
serial_number entry instead.

This CL also changes write-machine-info to always truncate the
machine-info file before generating contents. This is helpful
when developing and doing restarts of ui, to always have the right
content read by Chrome. Since Chrome gives precedence to the first
entry for a given key, changes are never taken into account when
appending to machine-info.

Manually verified the behavior in addition to emerge tests.

BUG=b:334109751
TEST=FEATURES=test emerge-${BOARD} chromeos-login

Change-Id: Ic1b534496254542734845a2ea8f35e2ed571af93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5450797
Commit-Queue: Yves Arrouye <drcrash@chromium.org>
Tested-by: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>