system_api: patchpanel: add network_id to DownstreamNetwork
This patch adds the network_id parameter to DownstreamNetwork messages
returned by patchpanel for GetDownstreamNetworkInfo DBus calls. The
DownstreamNetwork message is also added to the answer of
CreateTetheredNetwork and CreateLocalOnlyNetork DBus calls.

Currently this network_id parameter is not functional and is set to
fixed values depending on the type of requests (Tethering or LocalOnly).

A follow-up patch will expose the DownstreamNetwork information in
patchpanel-client through the caller callbacks.

Another follow-up patch will add in patchpanel the logic to assign
network_ids to network interfaces associated with DownstreamNetworks.

BUG=b:325124473
BUG=b:331679921
TEST=FEATURES=test emerge-$BOARD chromeos-base/system_api patchpanel

Change-Id: Ic57452523c19bbd129fcfa1174f7d2008d39b9d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5464986
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>