power: Add AllowZeroChargeReadOnAC Dynamic Machine Quirk
power_supply_info and dump_power_status have a long history of crashes
on Flex, dating back to 2020 [1]. Even after the work done in
b/287889424 and b/172211883, reven (the flex board) accounts for 28% [2]
of power_supply_info crashes and 5% [3] of dump_power_status crashes.

The crashes come from power_supply, which drops bogus zero readings from
the battery. We believe a significant amount of crashes can be
attributed to dead or dying batteries amongst the old devices we
support- a zero reading from these devices would be truthful, not
bogus. Just simply changing power_supply to stop ignoring zero readings
is not a good enough solution. http://b/316213327 is an EC fix that
tries to distinguish between faulty and truthful battery readings, which
would help alleviate this problem- but it only applies to devices with a
smart battery. However, many of our devices don't use a smart battery
driver and instead use the generic ACPI battery driver [4]. On these
devices, we cannot reasonably distinguish between bogus and correct zero
readings. So, we make the assumption that a device that is connected to
line power and reads out zero charge has a dead or dying battery. With
these assumptions in place, this change adds a dynamic machine quirk
that will detect if the device is using the generic ACPI battery driver,
activating AllowZeroChargeReadOnAC if so. This quirk will keep 0 battery
readings if the line power is connected, thus preventing
power_supply_info and dump_power_status from crashing.

Chromebooks, that is, non reven devices, do not activate machine quirks
in their overlay, so this fix will not affect them. Any OEM that tries
to bringup a device with the generic battery driver will need to
address bogus readings themselves.

[1]: http://b/188734337
[2]: http://crash/browse?q=product_name%3D%27ChromeOS%27+AND+expanded_custom_data.ChromeCrashProto.magic_signature_1.name%3D%27%5BAssert%5D+power_supply_info%3A+main%27
[3]: https://crash.corp.google.com/browse?q=product_name%3D%27ChromeOS%27+AND+expanded_custom_data.ChromeCrashProto.magic_signature_1.name%3D%27%5BAssert%5D+dump_power_status%3A+main%27
[4]: linux/drivers/acpi/battery.c

BUG=b:291920258
TEST=do
1) cros_workon_make --board=reven chromeos-base/power_manager --test
2) cros deploy --board=reven $DUT_IP  chromeos-config-bsp
   chromeos-config power_manager
where $DUT_IP is a device with a dead battery and the generic driver.
On DUT:
3) verify that this log appears
"AllowZeroChargeReadOnAC Pref set to enabled"
4) verify that power_supply_info and dump_power_status do not
crash.

Change-Id: I1f3460b2f02f96c0e1327b5932c38cf73527b6da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5454720
Reviewed-by: Jesse McGuire <jessemcguire@google.com>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Andre Braga <andrebraga@google.com>
Commit-Queue: Andre Braga <andrebraga@google.com>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>