mojo_service_manager: Support uid in ServiceEvent
This is a part of the security context migration (b/333324147).

Support uid to receive ServiceEvent. To keep supporting security
context, a legacy map is added to map security context to a set of
(uid, remote id). Each pair refers to a remote in
`receiver_set_[uid]`.
This is just a workaround during the migration.

BUG=b:333309396
TEST=FEATURES=test emerge-zork mojo_service_manager

Change-Id: I07a4e5992797e6e33c67057059845bf3a75abbc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5439145
Reviewed-by: Ben Scarlato <akhna@google.com>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Chung-sheng Wu <chungsheng@google.com>