vm_tools: vmc: bail out if read_line fails in y/n prompt
If the read from stdin fails, such as when the terminal is closed before
entering a full line, the read_line() function could return EIO, and the
unwrap() call would cause vmc to panic, creating a crash report. Treat
this the same as the user saying "no" to handle it cleanly.

BUG=None
TEST=close crosh while vmc is waiting for a y/n answer

Change-Id: I83ce8b8524719eafce56faa09c492a41175442ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5466006
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>