init: startup: Report pseudo-clobber with debug build
When we are in debug mode, we are not invoking clobber, lighly
cleaning stateful ourselves. Still we need to add a message in clobber
log that a [pseudo-]clobber happened.

In StatefulWipeTest, place the ".developer_mode" file as the right place
under "/mnt/stateful_partition" for consistency with the real code.
It requires to create the |stateful_mount| object with the proper
stateful argument, not the pseudo root.

BUG=b:172221451
TEST=Unit tests.

Change-Id: I58dcdb0ffe5a7d72269372eb3706b532642c85aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5466185
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>