cryptohome: Have auth session scheduling use task runner
This changes the auth session manager to always execute work (scheduled
via RunWhenAvailable) via the mount thread task runner. Currently when
you schedule a bunch of tasks the work effectively gets done
synchronously, in the sense that each task will just execute the next
task in the queue when it finishes, and when the queue is empty RunWhen
will just execute the code directly.

The technical problem with this kind of "mostly synchronous" code is
that it can produce a call stack that just grows deeper and deeper as
you pile tasks onto the top of the stack, chaining them forever until
all work is done. A practical problem is that this makes it harder to
test and validate that code is async safe, because it almost never runs
asynchronously.

This change unfortunately requires updating a lot of the userdataauth
unit tests because a lot of them assume that the operations will be run
synchronously. So I had to drop in "run until idle" calls to yield up
control and have all the async code run. My hope is that this mostly
confined to test code; I didn't have to fix an non-test code to work
this way. However, there is some risk that this kind of code is more
pervasive and we've introduced some potential hangs.

BUG=b:328665591
TEST=FEATURES=test emerge-$BOARD cryptohome
TEST=tast run $DUT cryptohome.*

Change-Id: Id03a59394eba2f0b8a9f3f8b56a81e3854f0837f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5484215
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Howard Yang <hcyang@google.com>