crash: Also collect decoded i915_error_state.
Currently, (compressed) i915_error_state is attached to the crash
report, but after decompression still needs to be decoded by
aubinator_error_decode in order to be used. Decode the i915_error_state
on device and attach the compressed result to the report as well.

BUG=b:328129103, b:327256299
TEST=cros_sdk cros_workon_make --board=hatch --test crash-reporter
TEST=chrome://crash and manually upload report to crash-staging
TEST=USE="asan fuzzer" emerge-hatch crash-reporter, cros_fuzz
--board=hatch shell, /usr/libexec/fuzzers/chrome_collector_fuzzer

Change-Id: I1a1e87974794fd75af7e5e7b094bd7223562d845
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5466922
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Jao-ke Chin-Lee <jchinlee@google.com>
Commit-Queue: Jao-ke Chin-Lee <jchinlee@google.com>
Reviewed-by: Aashay Shringarpure <aashay@google.com>