cryptohome: Fix a compile warning in migration helper unit tests
This fixes an annoying compilation warning. This code is flagged by the
compiler because it depends on VLAs, which is a C feature which is not
actually supported in proper C++ (although gcc and clang do kind of
support it as an extension).

In theory this could would work fine if GetParam() was a constexpr
function but it is not, and so instead I replaced the char[] with a
std::string which of course can handle dynamic sizing.

BUG=none
TEST=built unit tests, no more warning about a non-standard extension

Change-Id: I26fc1ad1591760d0a2c8e1875c5e305b0c6e2188
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5466920
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>