arc-setup: Classify (arcpp|arcvm)-specific deps in BUILD.gn
Some of the dependencies of the "all" target in arc/setup/BUILD.gn are
only needed for ARC++ container or ARCVM. Fix it to declare only
necessary dependencies by checking arcpp/arcvm USE flags.

Note that this CL causes the following minijail config file to be
installed only on ARC++ container:
* init/minijail/arc-boot-continue.conf
and the following ones to be installed only on ARCVM:
* init/minijail/arc-create-data.conf
* init/minijail/arc-handle-upgrade.conf
* init/minijail/arcvm-per-board-features.conf

The other files are unaffected, since the actual installation logics are
currently in the ebuild file.

BUG=b:335937499
TEST=`emerge-grunt arc-setup && \
     ls /build/grunt/usr/share/minijail | grep arc` shows:
     * arc-boot-continue.conf
     * arc-remote-data.conf
     * arc-stale-directory-remover.conf
TEST=`emerge-hatch arc-setup && \
     ls /build/hatch/usr/share/minijail | grep arc` shows:
     * arc-create-data.conf
     * arc-handle-upgrade.conf
     * arc-remote-data.conf
     * arc-stale-directory-remover.conf
     * arcvm-per-board-features.conf

Change-Id: I0e3674dbd7f0dbb88ba1ef7a2f4f4fce5d250333
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5467618
Reviewed-by: Satoshi Niwa <niwa@chromium.org>
Commit-Queue: Momoko Hattori <momohatt@chromium.org>
Tested-by: Momoko Hattori <momohatt@chromium.org>