cryptohome: update daemon-store permission if directory exists
Currently, if the user directory inside daemon-store has been created,
we won't re-create them or update the access permission. But that also
means when a daemon changes its configuration to update the permission,
it won't be propagated to the existing users directory.

Try to update the permission of the existing directories rather than
doing nothing.

BUG=b:327298508
TEST=Do the following manual check.
     (1) Sign in and take note of the permission of the directory
         /run/daemon-store/debugd/user_hash. Let's say it's 660.
     (2) chmod 123 /etc/daemon-store/debugd
     (3) reboot and sign in again
     (4) Verify that the permission of the directory in (1) is now
         updated to 123.
     (5) sign out and sign in as a guest, verify (4) again.

Change-Id: I5b3dc474f7c6d12487efdcc079c2a5b401b8af3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5472482
Tested-by: Archie Pusaka <apusaka@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: Archie Pusaka <apusaka@chromium.org>