fpstudy: Add Elan 80SG support to fpstudy.py
The Elan 80SG has a 14 bit pixel depth, which requires the use of 2 bytes
to represent. This script was designed to parse single byte raw values
for 8 bit depth sensors.

We add the Elan 80SG sensor parameters and expand the raw pixel
byte parser to be able to handle 2 bytes pixel values, when the
bits-per-pixel (bpp) is larger than 8.

Based on https://netpbm.sourceforge.net/doc/pgm.html, PGM format can
handle pixel depth up to 16 bits and can adjust the range based on
the specified max value in the PGM file. We already set the max value
in the PGM file, so the range is already correctly adjusted for 14 bit.
We simply need to expand the capabilities of the raw frame byte parser
to be able to read optionally 2 bytes per pixel.

BUG=b:331397247
TEST=None

Change-Id: Ife55a5ec633742497abd8eebfe8874d27df5764e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5478432
Commit-Queue: Craig Hesling <hesling@chromium.org>
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>