debugd: use D-Bus to retrieve sanitized username
brillo::cryptohome::home::GetDaemonStorePath() needs to read the system
salt, but that is not necessarily available when debugd starts. For
example it may not have been created on the first boot.
Instead, we use CryptohomeMiscInterfaceProxy::GetSanitizedUsername() to
retrieve the sanitized username that will be used to compute the path to
the daemon-store where the firmware dumps are stored.

BUG=b:336609164
TEST=cros_sdk FEATURES="test" emerge-${BOARD} debugd
TEST=cros deploy ${DUT} debugd
TEST=trigger 2 firmware dumps with
echo 1 > /sys/kernel/debug/iwlwifi/0000\:00\:14.3/iwlmvm/fw_dbg_collect
TEST=Call D-Bus GetFeedbackBinaryLogs method with collect_fw_dumps test
utility, verify the content of the created tarball matches the firmware
dumps

Change-Id: I6966dfdb068010af4f7ac48a1e8ff1b63dab8c25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5496372
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>