vm_tools: concierge: improve the response of the vm start failure
Currently, the failure reason in the dbus response of th VM start
failure contains only the VmStartChecker::Status int, which is hard to
spot the error location in the source code. This commit improves the
failure reason by making it align with the Concierge's log.

BUG=None
TEST=FEATURES=test emerge-betty-arc-t vm_host_tools

Change-Id: Ia6e4b90d9614d01fec3c7dc46a6e5867c25c4ca0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5483246
Commit-Queue: Takaya Saeki <takayas@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Tested-by: Takaya Saeki <takayas@chromium.org>