init: startup: Fix recovery logic
There was a logic error for testing when to trigger a recovery.
The code would dead loop in retries since the test !(uid != getuid())

Remove call to FileExists when we are going to check for uid, redundant.
Use pattern argument when backing up failed stateful.

BUG=none
TEST=Unit test

Fixes: ea5e0b45acb3 ("init: mount-encrypted: Use libstorage::Platform")
Change-Id: Ib2cbba5ad466fd6c0ea9dcf73c9f3e737ff97e72
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5484599
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>