smbprovider: Remove allow_credentials_update param
Remove allow_credentials_update attribute from
smbprovider/kerberos_artifact_synchronizer.cc, since it's currently
always set to true. This is a follow-up to the removal of Chromad
from smbprovider, which was completed in https://crrev.com/c/4518111.

The only scenario in which Kerberos credentials were not allowed to be
updated was on Chromad devices, and they are no longer supported.

BUG=b:263367348
TEST=cros_sdk USE=test emerge-${BOARD} smbprovider

Change-Id: I221c69c0445d51675ec1d3e07d8ffca6ee9778be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5485167
Reviewed-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Josh Simmons <simmonsjosh@google.com>
Commit-Queue: Felipe Andrade <fsandrade@chromium.org>
Tested-by: Felipe Andrade <fsandrade@chromium.org>