mini_udisks: Stop using the actual EFI system partition
The ESP on ChromeOS Flex is fairly small, so it is not a great place to
store UEFI capsule updates. Instead, capsules will be written to an
unencrypted directory on the stateful partition.

Drop all of the code to get information about the actual ESP, and
instead provide fake disk information. The mount point is now
`/run/uefi_capsule_updates`, which is a directory that will exist via
bind-mount inside the fwupd minijail sandbox.

BUG=b:328780545
TEST=cros_workon start chromeos-base/mini_udisks
TEST=FEATURES=test emerge-reven chromeos-base/mini_udisks

Change-Id: Ie6a53dffe6f404983d12dc433618ae9c95e2e2a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5502969
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>