system_api: Add flex_hwis shared header
Add a shared set of keys for Flex/reven for storing/naming our common
string format for hardware info. The strings referred to by these keys
are used across a number of not-really-user-facing surfaces for SWEs QE,
and Tpgm/PM-types to understand what hardware Flex is running on,
crashing on, getting feedback on, etc.

Add a header with constants for:
* Where we keep our hardware cache
* Key names for sending the data in common string:string key:value
  places, like crash reporter and feedback

BUG=b:300429988
TEST=CQ / future CLs

Change-Id: I95e07946f26fb7982d4d9fdd928e19657c50eb5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5502973
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Justin Ouellette <justinoue@google.com>
Commit-Queue: Ted Brandston <tbrandston@google.com>
Tested-by: Ted Brandston <tbrandston@google.com>