cryptohome: Add GetVaultProperties  call
This CL adds GetVaultProperties call which returns, as the name
suggests, the type of encryption used. Chrome will use this to display
the encryption in settings.

Something to note here is that this CL makes GetMountType functions
public because it was required for this CL, and seemed like the right
thing to do. Another approach could have been to use Homedirs functions,
but that seemed a little more brittle.

BUG=b:338187034
TEST=FEATURES=test emerge-zork cryptohome

Change-Id: Ieed3d4613556aad6a076e20b099289c6747ab855
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5505853
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Auto-Submit: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>