chromeos-dbus-bindings: Update white space handling.
Currently, white spaces and empty lines are handled manually.
On switching from the old tool written in C++ to this one,
we decided to keep the white spaces so that we can easily verify
the gap between the result of the old tool and the one of the new tool.

Now, we no longer need this, so this CL simplifies the white space
handling.
Note that, in the future we probably would like to apply code formatter
to the results. Though it's somehow nice to be kept for easier unittesting,
at least for now.

BUG=b:338002690
TEST=Tryjob

Change-Id: I8e20f01855d3ab2ad7ba8af5870ca5238435c0dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5508097
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Grace Cham <hscham@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>