missive: Declare scoped callback, use with weakptr
Ensure callbacks called even if WeakPtr is revoked

In code like
    void CallHandler(scoped_refptr<Handler> handler,
                     base::OnceCallback<void(ResultOrError)> cb) {
      handler->method(
          ...,
          base::BindOnce(&Handler::OnComplete,
	                 handler->GetWeakPtr(),
		         std::move(cb),
		         ...));
    }

`cb` is not called if `handler` is destructed and the weak pointer
is revoked. We need to ensure it is called even in this case.

This is done by wrapping `cb` in newly defined `Scoped` class.

Unit tests also added.

BUG=b:325905699
TEST=cros_sdk cros_workon_make --board=${BOARD} missive --test

Change-Id: I3664721810905fa8b78c1b162ef97b0f28338171
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5513677
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Alberto Juarez <albertojuarez@google.com>
Reviewed-by: Josh Hilke <jrhilke@google.com>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>