net-base: Don't format netlink payload if VLOG is off
As we are preparing the log payload outside of the VLOG statement, we
need to manually add an early return in case the VLOG is not going to be
printed anyway.

Saves around 400ms of CPU time on treeya360 boot.

BUG=b:338950691
TEST=flash CQ image, run perf at boot time, check that vsnprintf is out of the top stacks

Change-Id: I65e1829786bab35836a351f0bf88e5451d3679b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5514874
Auto-Submit: Tatsuyuki Ishi <ishitatsuyuki@google.com>
Reviewed-by: Cassie Wang <cassiewang@google.com>
Commit-Queue: Tatsuyuki Ishi <ishitatsuyuki@google.com>
Reviewed-by: Jason Iman <jasongustaman@chromium.org>
Tested-by: Tatsuyuki Ishi <ishitatsuyuki@google.com>