biod: Add metrics for operation return status
Add metrics reporting for the [Operation]Reply::[Operation]Status fields
we return as dbus responses to the operations. Put the reporting logic
in wrapper component so that the auth stack manager doesn't have to both
set the response status field and report the metrics in each branch.

BUG=b:253993586
TEST=biod unittests

Change-Id: Id0276bb8d3e03adb5f95dcd366bad432b5d4d64c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5518105
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Patryk Duda <patrykd@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>