biod: Reset context when uploading template
We need to reset context and clear the existing templates (if any) in
FPMCU before uploading a new set of template. Also, keep the loaded user
state when the user logs out so that we're consistent with the FPMCU set
of template. We decide not to clear the user state from both biod and
FPMCU when a user logs out because in most of the cases, the same user
will instantly be loaded back again.

These changes ensure that the set of records in biod and in FPMCU are
now always consistent.

BUG=b:339162287
TEST=biod unittests
TEST=manual test that after deleting a template, the rest of the
templates are consistent with FPMCU and match succeeds (with the
remaining templates but not the deleted).

Disallow-Recycled-Builds: test-failures
Change-Id: If3974a9de6a798c6cb870e5eb3f5f59224c70e59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5514414
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Patryk Duda <patrykd@google.com>