vm_tools: concierge: make vmm context-types selection explicit
The logic is confusing. If no context types are set, then the VMM
internally chooses which to advertise to the VM userspace. If any are
explicitly set, inference is disabled and only those explicitly enabled
are advertised.

Try to avoid confusion by always being explicit when building the
StartVm arguments. If inference behavior ever changes in the vmm, we are
insulated from negative consequences.

BUG=b:338665678
TEST=cros_sdk emerge-guybrush chromeos-base/vm_host_tools
TEST=ARCVM works as expected on guybrush
TEST=ARCVM (with USE=cross_domain_context) also works on guybrush
TEST=Borealis works as expected on guybrush

Change-Id: Ie55551c575ca91588f6b4656b083ab6671c46ebb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5522935
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Ryan Neph <ryanneph@google.com>
Commit-Queue: Ryan Neph <ryanneph@google.com>
Reviewed-by: Yiwei Zhang <zzyiwei@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>