vm_tools: sommelier: WaylandChannel early-init with fallback
While migrating from VirtWaylandChannel to VirtGpuChannel
(cross-domain), sommelier should attempt to fallback to
VirtWaylandChannel in case the host VMM doesn't provide a Virtio-Gpu
cross-domain context-type implementation.

This way, VMs may opt-in to VirtGpuChannel when it is possible, and
continue to operate normally on VirtWaylandChannel when it is not.

Also refactors the event loop initialization flow that depends on the
wayland channel.

BUG=b:314788008
TEST=`sommelier --virtgpu-channel ...` falls back to VirtWaylandChannel
when Virtio-Gpu doesn't advertise its "cross-domain" context-type.
TEST=USE=fuzzer cros_sdk emerge-tatl sommelier
TEST=cros_sdk cros_workon_make --verbose --board tatl --test sommelier

Change-Id: I925bef4af41dc5fa76a6816191b15b34b27384f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5523855
Reviewed-by: Chloe Pelling <cpelling@google.com>
Commit-Queue: Ryan Neph <ryanneph@google.com>
Tested-by: Ryan Neph <ryanneph@google.com>