libtouchraw: Allow consumers to modify heatmap
Heatmaps are the final output of libtouchraw and ownership is passed on
to consumers. It is therefore expected that consumers once they have
ownership should be able to modify it. Having heatmap be const will
require consumers to either copy or perform a const_cast before they
can modify or operate on the heatmap. This change removes the const
to allow consumers to modify the heatmap if they need to.

BUG=b:332429708, b:331432877
TEST=CQ

Change-Id: I12ca96c8a3c90acad74919abe5a3d841ed14f0e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5523864
Reviewed-by: Jingyuan Liang <jingyliang@chromium.org>
Tested-by: Henry Barnor <hbarnor@chromium.org>
Reviewed-by: William Mahon <wmahon@google.com>
Commit-Queue: Henry Barnor <hbarnor@chromium.org>
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Xinglong Luan <alanlxl@chromium.org>