arc: libvda: handle RequestVideoFrames response
This CL updates the VideoFramePoolClient mojo interface used by the
direct VD path in Chrome. CL:5521961 added a callback for the
RequestVideoFrames function so Chrome can know if the frames it is
receiving are from the last initialization of the pool. This solves a
bug in the arc.VideoDecodeAccelVD.*_vm Tast tests.

This response callback is supposed to be analogous to the
AssignPictureBuffers() call of the VDA interface, so the time at which
the callback is called is the same.

BUG=b:221768154
TEST=tast run ${IP} arc.VideoDecodeAccelVD.h264_vm (on hatch)

Change-Id: Id71805d2be0df6d96ac35fe8e0604fc94d4dc272
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5521459
Commit-Queue: Pilar Molina Lopez <pmolinalopez@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@google.com>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Tested-by: Pilar Molina Lopez <pmolinalopez@chromium.org>