shill: Add OnDeviceClaimed to device objects
Add a new OnDeviceClaimed method on device objects in Shill. The base
version of this method does nothing, but subclasses can use it to
perform special logic when their interface is claimed. This will be used
by WiFi device to remove the interface from wpa_supplicant when the
interface is claimed.

BUG=b:328073329
TEST=FEATURES=test emerge-brya shill
TEST=dbus-send --system --fixed --print-reply --dest=org.chromium.flimflam / org.chromium.flimflam.Manager.ClaimInterface string:claimer string:wlan0
2024-05-08T21:48:07.743605Z DEBUG shill[31167]: VERBOSE2 shill: [device.cc(762)] /device/wlan0 OnDeviceClaimed

Change-Id: I326424f8f2d937c979422da15d086011573fe9ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5530402
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>