shill: Be more selective about removing wifi interface from supplicant
Currently, we call RemoveInterface on wpa_supplicant whenever the Shill
WiFi device object is disabled. With upcoming concurrency work, we will
be supporting use-cases where we need wpa_supplicant to maintain it's
interface reference while the Shill object is disabled. So, stop calling
RemoveInterface in WiFi::Stop().

The reason this logic was added in the first place was to support the
ClaimInterface command, which allows an outside process to claim
ownership of an interface and prevent other processes (e.x. Shill and
wpasupplicant) from using it. So, we maintain support for this command
by implementing WiFi::OnDeviceClaimed() and calling RemoveInterface from
there. This way, the device still correctly gets removed from
wpa_supplicant in the device claim flow, but stays available in
supplicant during the normal disable flow.

BUG=b:328073329
TEST=FEATURES=test emerge-brya shill
TEST=Supplicant keeps tracking the device when WiFi is enabled/disabled
in shill:
http://gpaste/6574431413731328
TEST=Supplicant stops tracking the device when the interface is claimed:
http://gpaste/5559808136839168

Change-Id: Ib4eedb04f0a04ed94d51cfd433e12fbc1c2b3f2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5530403
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>