cryptohome: Avoid DeletePathRecursively when migrating Downloads folder
When removing the old ~/Downloads and ~/Downloads-backup folders after a
Downloads folder migration, use DeleteFile() instead of
DeletePathRecursively(). These old folders should be empty, and
DeleteFile() should be able to remove them. If they're not empty, then
something unexpected happened, and we don't want to take the risk of
inadvertently erasing some actual data. The existing UMA histograms
should also inform us if this is happening.

Similarly, when migrating the contents of the existing Downloads mount
point, do not erase an item if it cannot be moved. First, the Downloads
mount point should be empty, and we don't really expect to have to move
anything. If there is something to move from there, something unexpected
happened. Second, if an item cannot be moved, now that we have a proper
name collision resolution system, then something really unexpected
happened, and we don't want to take the risk of inadvertently erasing
some actual data. A new UMA histogram will inform us if this is
happening.

BUG=b:339527323
TEST=P2_TEST_FILTER="DownloadsBindMountMigrationTest.*" cros_workon_make --board=$BOARD --test chromeos-base/cryptohome

Change-Id: I411b8ef8fbe642ba7944abca03c01129ae3958b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5528333
Commit-Queue: François Degros <fdegros@chromium.org>
Reviewed-by: Wenbo Jie <wenbojie@chromium.org>
Tested-by: François Degros <fdegros@chromium.org>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>