chromeos-dbus-binding: Implement adapter method generation.
This CL introduces the new adapter method generation.
Currently, all message handling is done in some complex callback stack,
so it was difficult to find, e.g., what method is failing if
read argument is something unexpected.

With this change, we expect the method name will be appeard
in the stack trace on crashing.

Note: there seems some issue (UB) on generated code, but because of
the bot dependency issue, it looks like difficult to run UB-san
without landing the change. To break the circular dependency,
this CL lands the feature to generate the code, but guarded
under the flag.

BUG=b/289932268
TEST=Tryjob

Change-Id: I49b15e0ad2a1be34774a0acf1839151fb74bc623
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5584441
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Grace Cham <hscham@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>