typecd: Add captive cable options to speed metric
The cable speed metric in typecd will define a cable speed based
entirely on the cable's properties. That makes typecd unable to
determine when a captive cable is connected. Often, captive cables are
reported by the partner. This updates typecd to consider a captive cable
condition for non-emarked cables when determining the cable speed.

BUG=b:338471345
TEST=FEATURES=${BOARD} emerge-brya chromeos-base/typecd

Change-Id: Ib3c475e53dd04023c6056d01c99b7c09395ac41c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5529228
Reviewed-by: Benson Leung <bleung@google.com>
Commit-Queue: Jameson Thies <jthies@google.com>
Tested-by: Jameson Thies <jthies@google.com>